refactor: PostDetailPageからCommentSectionコンポーネントを分離
PostDetailPage: 198行→94行（104行削減）
CommentSection: コメントヘッダー・入力フォーム・一覧・返信フォーム・ネスト返信を管理

closes #1393
